refactor(analysis): split nm_tool_stats into focused modules
Extract four focused modules from nm_tool_stats.py (~2300 lines):
- nm_stat_func.py: NMStatFunc hierarchy, badvalue(), FUNC_NAMES_* constants
- nm_stat_utilities.py: stat(), stats(), find_level_crossings(), linear_regression()
- nm_stat_win.py: NMStatWin, NMStatWinContainer
- nm_tool_stats.py: trimmed to NMToolStats and NMToolStats2 only

Split test_nm_stats.py into matching test modules:
- test_nm_stat_func.py, test_nm_stat_utilities.py, test_nm_stat_win.py
- test_nm_tool_stats.py, test_nm_stat_igor.py

No behaviour changes. 1200 tests passing.
